+++ src/panoptes/utils/config/server.py
@@ -7,9 +7,9 @@ from flask import jsonify
 from flask import request
 from flask.json import JSONEncoder
 from gevent.pywsgi import WSGIServer
+from loguru import logger
 from panoptes.utils.config.helpers import load_config
 from panoptes.utils.config.helpers import save_config
-from panoptes.utils.logging import logger
 from panoptes.utils.serializers import serialize_object
 from scalpl import Cut
