+++ src/panoptes/utils/error.py
@@ -1,6 +1,6 @@
 import sys
 
-from panoptes.utils.logging import logger
+from loguru import logger
 
 
 class PanError(Exception):
