+++ src/panoptes/utils/social/slack.py
@@ -1,5 +1,5 @@
 import requests
-from panoptes.utils.logging import logger
+from loguru import logger
 
 
 class SocialSlack(object):
